SqlStorage: - удалил поле  ConnectionFactory connectionFactory - добавил поле SqlHelper helper - изменил коструктор - удалил вложенный класс SqlStorage - поправил остальной код согласно изменениям выше
добавил и реализовал отдельный класс SqlStorage
