+++ src/ru/javawebinar/basejava/sql/SqlStorage.java
@@ -2,7 +2,6 @@ package ru.javawebinar.basejava.sql;
 
 import ru.javawebinar.basejava.exception.ExistStorageException;
 import ru.javawebinar.basejava.exception.NotExistStorageException;
-import ru.javawebinar.basejava.exception.StorageException;
 import ru.javawebinar.basejava.model.Resume;
 import ru.javawebinar.basejava.storage.Storage;
 
@@ -11,15 +10,16 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class SqlStorage implements Storage {
-    public ConnectionFactory connectionFactory;
+    private final SqlHelper helper;
+
 
     public SqlStorage(String dbUrl, String dbUser, String dbPassword) {
-        connectionFactory = () -> DriverManager.getConnection(dbUrl, dbUser, dbPassword);
+        helper = new SqlHelper(() -> DriverManager.getConnection(dbUrl, dbUser, dbPassword));
     }
 
     @Override
     public void clear() {
-        new SqlHelper(connectionFactory).startExecution("DELETE FROM resume", (Executor<Void>) ps -> {
+        helper.startExecution("DELETE FROM resume", (Executor<Void>) ps -> {
             ps.execute();
             return null;
         });
@@ -27,7 +27,7 @@ public class SqlStorage implements Storage {
 
     @Override
     public void update(Resume r) {
-        new SqlHelper(connectionFactory).startExecution("UPDATE resume SET full_name=? WHERE uuid=?", (Executor<Void>) ps -> {
+        helper.startExecution("UPDATE resume SET full_name=? WHERE uuid=?", (Executor<Void>) ps -> {
             ps.setString(1, r.getFullName());
             ps.setString(2, r.getUuid());
             executeUpdate(ps, r.getUuid());
@@ -37,7 +37,7 @@ public class SqlStorage implements Storage {
 
     @Override
     public void save(Resume r) {
-        new SqlHelper(connectionFactory).startExecution("INSERT INTO resume(uuid, full_name) VALUES (?,?)", (Executor<Void>) ps -> {
+        helper.startExecution("INSERT INTO resume(uuid, full_name) VALUES (?,?)", (Executor<Void>) ps -> {
             ps.setString(1, r.getUuid());
             ps.setString(2, r.getFullName());
             try {
@@ -53,7 +53,7 @@ public class SqlStorage implements Storage {
 
     @Override
     public Resume get(String uuid) {
-        return new SqlHelper(connectionFactory).startExecution("SELECT * FROM resume WHERE uuid =?", ps -> {
+        return helper.startExecution("SELECT * FROM resume WHERE uuid =?", ps -> {
             ps.setString(1, uuid);
             ResultSet rs = ps.executeQuery();
             if (!rs.next()) {
@@ -65,7 +65,7 @@ public class SqlStorage implements Storage {
 
     @Override
     public void delete(String uuid) {
-        new SqlHelper(connectionFactory).startExecution("DELETE FROM resume WHERE uuid=?", ps -> {
+        helper.startExecution("DELETE FROM resume WHERE uuid=?", ps -> {
             ps.setString(1, uuid);
             executeUpdate(ps, uuid);
             return null;
@@ -75,7 +75,7 @@ public class SqlStorage implements Storage {
     @Override
     public List<Resume> getAllSorted() {
         List<Resume> result = new ArrayList<>();
-        new SqlHelper(connectionFactory).startExecution("SELECT * FROM resume ORDER BY full_name, uuid", (Executor<Void>) ps -> {
+        helper.startExecution("SELECT * FROM resume ORDER BY full_name, uuid", (Executor<Void>) ps -> {
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
                 result.add(new Resume(rs.getString("uuid"), rs.getString("full_name")));
@@ -87,7 +87,7 @@ public class SqlStorage implements Storage {
 
     @Override
     public int size() {
-        return new SqlHelper(connectionFactory).startExecution("SELECT COUNT(*) FROM resume", ps -> {
+        return helper.startExecution("SELECT COUNT(*) FROM resume", ps -> {
             ResultSet rs = ps.executeQuery();
             rs.next();
             return rs.getInt("count");
@@ -97,23 +97,6 @@ public class SqlStorage implements Storage {
     private void executeUpdate(PreparedStatement ps, String uuid) throws SQLException {
         if (ps.executeUpdate() == 0) {
             throw new NotExistStorageException(uuid);
-        }
-    }
-
-    public static class SqlHelper {
-        private final ConnectionFactory connection;
-
-        public SqlHelper(ConnectionFactory connection) {
-            this.connection = connection;
-        }
-
-        public <T> T startExecution(String string, Executor<T> executor) {
-            try (Connection conn = connection.getConnection();
-                 PreparedStatement ps = conn.prepareStatement(string)) {
-                return executor.execute(ps);
-            } catch (SQLException e) {
-                throw new StorageException(e);
-            }
         }
     }
 }
